docker/vessel_tracker/requirements.txt
boto3==1.34.19
botocore==1.34.19
dnspython==2.4.2
jmespath==1.0.1
paho-mqtt==1.6.1
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.10.0
six==1.16.0
urllib3==1.26.18
